Split app.py into models, forms, views
