+++ ca/models.py
@@ -1,0 +1,20 @@
+from ca import app, db
+
+
+class Request(db.Model):
+    id = db.Column(db.String(80), primary_key=True)
+    email = db.Column(db.String(120))
+    generation_date = db.Column(db.Date())
+
+    def __init__(self, id, email, generation_date=None):
+        self.id = id
+        self.email = email
+        self.generation_date = generation_date
+
+    def __repr__(self):
+        return "<Request {} - {} - {}>".format(
+                self.id,
+                self.email,
+                self.generation_date
+                )
+
