+++ ca/static/lib/bootstrap/css/bootstrap-theme.css
