+++ ca/static/lib/bootstrap/fonts/glyphicons-halflings-regular.ttf
