+++ ca/static/lib/bootstrap/js/bootstrap.js
