+++ manage.py
@@ -3,7 +3,7 @@
 
 from flask.ext.script import Manager
 from flask.ext.migrate import Migrate, MigrateCommand
-from app import app, db
+from ca import app, db
 
 migrate = Migrate(app, db)
